Hard revert back to 2.8.0 tagged release
I ended up breaking things.
Kept a couple of simple test files I was using. They're basically empty shells anyway.
And added the github workflow for building the .curapackage
