+++ .github/workflows/package.json
@@ -1,0 +1,16 @@
+{
+    "author": {
+        "author_id": "Slashee-the-Cow",
+        "display_name": "Slashee the Cow",
+        "email": "slasheethecow@gmail.com",
+        "website": "http://github.com/Slashee-the-Cow/"
+    },
+    "description": "Originally by 5@xes - This plugin adds several custom supports you can use for your model instead of the support automatically added by Cura.\n\nGiven new life by Slashee the Cow so everyone can use this awesome tool.",
+    "display_name": "Calibration Shapes Reborn",
+    "package_id": "CustomSupportsReborn",
+    "package_type": "plugin",
+    "package_version": "1.0.0",
+    "sdk_version": 8,
+    "sdk_version_semver": "8.0.0",
+    "website": "https://github.com/Slashee-the-Cow/CustomSupportsReborn/"
+}
