+++ CustomSupportsRebornSimple.py
@@ -1,0 +1,5 @@
+from UM.Tool import Tool
+
+class CustomSupportsRebornSimple(Tool):
+    def __init__(self):
+        super().__init__()
