+++ MinimalTest.qml
@@ -1,0 +1,12 @@
+import QtQuick 6.2
+import QtQuick.Controls 6.2
+
+Rectangle {
+    width: 200
+    height: 100
+    color: "blue"
+    Text {
+        text: "Minimal Test Panel"
+        anchors.centerIn: parent
+    }
+}
