+++ plugin.json
@@ -1,9 +1,9 @@
 {
-    "name": "Custom Supports Cylinder",
-    "author": "5@xes",
-    "version": "2.8.2",
-    "description": "Create 6 types of custom supports",
-    "i18n-catalog": "customsupport",
-    "api": 7,
-    "supported_sdk_versions": ["7.0.0", "7.1.0", "7.2.0", "7.3.0", "7.4.0", "7.5.0", "7.6.0", "7.7.0", "7.8.0", "7.9.0", "8.0.0", "8.1.0", "8.2.0"]
+    "name": "Custom Supports Reborn",
+    "author": "Slashee the Cow",
+    "version": "1.0.0",
+    "description": "Originally by 5axes - Gives you several different kinds of custom support to add to your scene.",
+    "i18n-catalog": "customsupportsreborn",
+    "minimum_cura_version": "5.0",
+    "api": 8
 }
