--- resources/i18n/fr_FR/customsupport.po
@@ -1,161 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the Custom Support Cylinder package.
-# 5axes <5axes@free.fr>, 2023.
-msgid ""
-msgstr ""
-"Project-Id-Version: Custom Support Cylinder\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-03-14 02:34+0000\n"
-"PO-Revision-Date: 2023-03-14 03:35+0200\n"
-"Last-Translator: 5axes <5axes@free.fr>\n"
-"Language-Team: 5axes\n"
-"Language: fr\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n > 1);\n"
-"X-Generator: Virtaal 0.7.1\n"
-
-#: CustomSupportsCylinder.py:130 CustomSupportsCylinder.py:1214
-msgctxt "@label"
-msgid "Remove All"
-msgstr "Supprimer Tous"
-
-#: CustomSupportsCylinder.py:368
-msgctxt "@info:label"
-msgid "Info modification current profile '"
-msgstr "Information modification du profil actuel Paramètre '"
-
-#: CustomSupportsCylinder.py:368
-msgctxt "@info:label"
-msgid ""
-"' parameter\n"
-"New value : "
-msgstr ""
-"'\n"
-"Nouvelle Valeur : "
-
-#: CustomSupportsCylinder.py:368
-msgctxt "@info:label"
-msgid "Skirt"
-msgstr "Jupe"
-
-#: CustomSupportsCylinder.py:369 CustomSupportsCylinder.py:484
-msgctxt "@info:title"
-msgid "Warning ! Custom Supports Cylinder"
-msgstr "Attention ! Supports Custom Cylindrique"
-
-#: CustomSupportsCylinder.py:484
-msgctxt "@info:label"
-msgid "Info modification support_type new value : Everywhere"
-msgstr "Info modification support_type nouvelle valeur : Partout"
-
-#: CustomSupportsCylinder.py:496
-msgctxt "@label"
-msgid "Remove Last"
-msgstr "Supprimer Derniers"
-
-#: __init__.py:23
-msgctxt "@label"
-msgid "Custom Supports Cylinder"
-msgstr "Supports Custom Cylindrique"
-
-#: __init__.py:24
-msgctxt "@info:tooltip"
-msgid "Add 6 types of custom support"
-msgstr "Ajouter 6 types de support personnalisé"
-
-#: qml/qml_qt5/CustomSupport.qml:60 qml/qml_qt6/CustomSupport.qml:63
-msgctxt "@label"
-msgid "Cylinder"
-msgstr "Cylindre"
-
-#: qml/qml_qt5/CustomSupport.qml:73 qml/qml_qt6/CustomSupport.qml:79
-msgctxt "@label"
-msgid "Tube"
-msgstr "Tube"
-
-#: qml/qml_qt5/CustomSupport.qml:86 qml/qml_qt6/CustomSupport.qml:95
-msgctxt "@label"
-msgid "Cube"
-msgstr "Cube"
-
-#: qml/qml_qt5/CustomSupport.qml:99 qml/qml_qt6/CustomSupport.qml:119
-msgctxt "@label"
-msgid "Abutment"
-msgstr "Contrefort"
-
-#: qml/qml_qt5/CustomSupport.qml:112 qml/qml_qt6/CustomSupport.qml:135
-msgctxt "@label"
-msgid "Freeform"
-msgstr "Forme libre"
-
-#: qml/qml_qt5/CustomSupport.qml:125 qml/qml_qt6/CustomSupport.qml:151
-msgctxt "@label"
-msgid "Custom"
-msgstr "Personnalisé"
-
-#: qml/qml_qt5/CustomSupport.qml:150 qml/qml_qt6/CustomSupport.qml:180
-msgctxt "@label"
-msgid "Size"
-msgstr "Taille"
-
-#: qml/qml_qt5/CustomSupport.qml:161 qml/qml_qt6/CustomSupport.qml:191
-msgctxt "@label"
-msgid "Max Size"
-msgstr "Taille maxi"
-
-#: qml/qml_qt5/CustomSupport.qml:173 qml/qml_qt6/CustomSupport.qml:203
-msgctxt "@label"
-msgid "Type"
-msgstr "Type"
-
-#: qml/qml_qt5/CustomSupport.qml:185 qml/qml_qt6/CustomSupport.qml:216
-msgctxt "@label"
-msgid "Interior Size"
-msgstr "Taille intérieure"
-
-#: qml/qml_qt5/CustomSupport.qml:197 qml/qml_qt6/CustomSupport.qml:228
-msgctxt "@label"
-msgid "Angle"
-msgstr "Angle"
-
-#: qml/qml_qt5/CustomSupport.qml:339 qml/qml_qt6/CustomSupport.qml:367
-msgctxt "@option:check"
-msgid "Define as Model"
-msgstr "Définir comme Modèle"
-
-#: qml/qml_qt5/CustomSupport.qml:353 qml/qml_qt6/CustomSupport.qml:380
-msgctxt "@option:check"
-msgid "Set on Y direction"
-msgstr "Orienter selon Y"
-
-#: qml/qml_qt5/CustomSupport.qml:353 qml/qml_qt6/CustomSupport.qml:380
-msgctxt "@option:check"
-msgid "Set on Main direction"
-msgstr "Orienter direction principale"
-
-#: qml/qml_qt5/CustomSupport.qml:367 qml/qml_qt6/CustomSupport.qml:393
-msgctxt "@option:check"
-msgid "Rotate 180°"
-msgstr "Rotation 180°"
-
-#: qml/qml_qt5/CustomSupport.qml:381 qml/qml_qt6/CustomSupport.qml:406
-msgctxt "@option:check"
-msgid "Auto Orientation"
-msgstr "Orientation automatique"
-
-#: qml/qml_qt5/CustomSupport.qml:395 qml/qml_qt6/CustomSupport.qml:419
-msgctxt "@option:check"
-msgid "Scaling in main Directions"
-msgstr "Echelle selon directions principales"
-
-#: qml/qml_qt5/CustomSupport.qml:409 qml/qml_qt6/CustomSupport.qml:431
-msgctxt "@option:check"
-msgid "Equalize heights"
-msgstr "Égaliser les hauteurs"
-
-#~ msgctxt "@option:check"
-#~ msgid "Define as Part"
-#~ msgstr "Définir comme pièce"
